fix tons of bugs

+++ src/main/java/me/alegian/thavma/impl/common/menu/slot/DynamicSlot.kt
@@ -6,9 +6,9 @@ import me.alegian.thavma.impl.common.menu.Menu
  * Used for adding size to slots, allowing them to be dynamically positioned
  */
 interface DynamicSlot<T : Menu> {
-  var actualX: Int
+  var actualX: Float
 
-  var actualY: Int
+  var actualY: Float
 
   var size: Int
